Gallery of Samurai Art
The project is an interactive landing page focused on showcasing Japanese Sumerai art through a modern and elegant design. The site was developed using HTML5 and CSS to design the interface, with JavaScript added to add seamless interactions such as navigating between sections and viewing the image gallery using the Swiper.js library. The project is ideal for showcasing artwork or projects that rely on visual imagery.